+++ bin/sb/src/commands/mod.rs
@@ -2,7 +2,6 @@ pub mod describe;
 pub mod eval;
 pub mod logs;
 pub mod query;
-pub mod repl;
 pub mod script;
 pub mod space;
 pub mod upgrade;